services/ai-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
openai==1.35.0
google-generativeai==0.5.4
httpx==0.27.0
